clean up

+++ src/main/java/ua/epam/payments/payments/model/dao/mapper/UserMapper.java
@@ -16,7 +16,7 @@ public class UserMapper {
                 .withEmail(rs.getString(5))
                 .withPassword(rs.getString(6))
                 .withBlocked(rs.getBoolean(7))
-                .withRolesId(rs.getLong(8))
+                .withRolesId(rs.getInt(8))
                 .build();
 
     }
